requirements.txt
discord.py 
web3 
python-dotenv 
requests 
